Create RowEdit component and update usage.
